libct/error.go: rm ConfigError, add ErrInvalidConfig
Replace ConfigError type with ErrInvalidConfig variable.

This is somewhat complicated, as we need to make all errors returned
from config validator to be ErrInvalidConfig, while keeping them
unwrappable.  This is implemented by introducing a type which holds an
underlying error, and adding Is method that answers that the error is
ErrInvalidConfig.

[v2: silence the linter]

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>