libct/cg/sd/v1: Set: don't leave cgroup frozen
I. In case a parent cgroup is frozen, the current code reads the cgroup
   state (into targetFreezerState) as FROZEN, and sets it explicitly after
   setting unit properties. This is obviously wrong, as we should not
   explicitly freeze the cgroup because its parent was frozen.

   The issue happens because:

   1. The m.GetFreezerState method does not distinguish between
      "self frozen" and "parent frozen" states (i.e. it does not
      consult freezer.self_freezing).

   2. The m.Freeze method changes m.cgroups.Resources.Frozen field.

II. The current code does freeze/thaw unconditionally, meaning it
    freezes the already frozen cgroup (which is a no-op except it
    still requires some reads and writes to freezer.state), and
    thaws the cgroup which is about to be frozen.

Solve all this by figuring out whether we need to freeze and thaw,
and introducing and using a method which does not change value of
m.cgroups.Resources.Frozen.

NOTE the current code assume that the container that is currently frozen
will remain frozen for the duration of SetUnitProperties. This might not
be true but there's no way to guarantee it (even in case we freeze the
cgroup ourselves).

Reported-by: Odin Ugedal <odin@uged.al>
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>