libct/cg/v1: workaround CPU quota period set failure
As reported in issue 3084, sometimes setting CPU quota period fails
when a new period is lower and a parent cgroup has CPU quota limit set.

This happens as in cgroup v1 the quota and the period can not be set
together (this is fixed in v2), and since the period is being set first,
new_limit = old_quota/new_period may be higher than the parent cgroup
limit.

The fix is to retry setting the period after the quota, to cover all
possible scenarios.

Add a test case to avoid regressions. Now, this is somewhat complicated
as there's no easy way to create a transient unit, so a binary,
sd-helper, had to be added. On top of that, an ability to create a
parent/pod cgroup is added, which might be useful for future integration
tests.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>