libct/logs: test: make more robust
When playing with the log forwarder, I broke it, but all the units tests
were still passing. This happened because test cases were merely looking
for a work like "kitten" in the log output, which was also there as a
part of an error message produced by log forwarder.

Make the test a bit more robust by
 - looking for a complete log message, not just part of it;
 - also checking that log file does not contain errors.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>