libct/cg/sd/v1: fix freezeBeforeSet
Commit f2db87986cab5a7c9 was not working as expected:

1. It did not do the right job getting unit properties;
2. The check that deviceAllow is empty was not working.

While there is a way to fix both issues, let's employ a simpler check --
if r.SkipDevices is set, and devices.list only contains "allow all" rule,
we can safely skip freezing.

The only issue with the patch, in case we call freezeBeforeSet before
Apply(), the m.paths are not yet initialized, so we have to manually
figure the path to devices subsystem in haveDeviceRules. This is being
fixed elsewhere.

Fixes: f2db879
Reported-by: Odin Ugedal <odin@uged.al>
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>