libct/cg/sd/v1: add a knob to skip freeze on set
Commit b810da149008f added support for Device* properties into systemd
cgroup v1 controller. Unfortunately, systemd re-applies device rules on
every resource update, and does so disruptively, which results in
spurious device permission errors. To work around that, systemd cgroup
v1 controller freezes the cgroup before (and thaws it after) applying
systemd properties.

This freeze fixes the issue for runc, but creates a problem for
kubernetes, which uses libcontainer/cgroups/systemd to manage pod slice
units. When kubelet updates a pod slice, all the containers under that
pod are unexpectedly frozen for some time, which leads to other issues.

One way to fix this was tried in commit f2db87986cab5a7c. By looking
into unit properties, we can predict whether systemd is going to update
device rules, and skip the freeze if it won't.

Unfortunately, the above commit contains a few bugs, resulting in
broken device rules check, so it's still not fixing the issue.

Upon fixing those bug (see PR 3143), a benchmark was created to assess
how expensive getting unit properties is. Turns out, it is -- Set()
times went from 200us to 500us (see full benchmark results in that PR).

An alternative approach was tried in PR 3151, in which we check
devices.list to find out whether they have some rules or just
a single "allow-all". Alas, this approach is not working if a cgroup
there is a parent with some device rules set.

Considering all this, it seems that the best approach is to add a
"do not freeze on set" flag, which kubernetes can use when it knows
the slice did not and does not have device rules, so freeze can safely
be skipped.

Remove the (expensive and currently broken) unit properties querying,
and add/use a SkipFreezeOnSet flag instead.

Since the logic is trivial, no tests are needed.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>