libct/nsenter: Don't ignore write failures
If write fails for some reason or manages to write less than what we
requested (e.g. disk is full), report an error.

This silence the warning when compiling on github CI, as now the
write(2) return code is not ignored.

While we could do:
	ret = write(.., ret)
	if (ret < 0)

as other parts of the code does, there was recently a race condition
fixed in "libct/nsenter: fix logging race in nsexec" (2bab4a56f) and
seems fragile to just test for -1. Let's check that we wrote exactly
what we expect, otherwise it won't be valid json written.

For the same reason, the race condition fixed, we don't retry if we
write partial data, as that uses two different write(2) calls which
might hit the same race. Therefore, if we failed to write what we
expect, we just let the caller know.

Note that if the write fails, it is possible for the (partial) json
written to not be valid.

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>