cpuset: option for updating cpus and mems in child groups
Trying to set cpuset.cpus of a cgroup so that cpus of a child is not a
subset of cpus of the cgroup causes an error (device or resource
busy).

New option --child-policy="copy" sets cpuset.cpus of the group and all
its children to the same value. Introducing a string-valued policy for
this purpose leaves room for alternative policies in the future. Other
sensible policies include, but are not limited to, modifying only
children that already share the same values with the group, and
keeping the relative number and intersections of cpus unchanged in all
child groups.

cpuset.mems is handled similarly to cpuset.cpus.

Signed-off-by: Antti Kervinen <antti.kervinen@intel.com>