script/release.sh: add arm64 cross-build
Modify the script to supply more architectures in addition to the native
build. Currently only arm64 is supported, but the code is ready for
adding more, if needed.

The additional arch is enabled by specifying "-a arm64" argument to
release.sh (or "make RELEASE_ARGS="-a arm64" release" if called via
make).

Modify the validate/release CI job to build for arm64 as well.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>