ChromeOS: correct paths to cpuset.cpus and cpuset.mems
My understanding is that between cgroups v1 and v2 paths for `cpus` and
`mems` files changed to include `cpuset.` prefix, but ChromeOS uses a
"legacy" mode of cgroups v2 where the old file names are used.

This patch is only intended for running runc on ChromeOS (in a chroot
managed by Crouton).