ChromeOS: chroot to the desired dir after entering mnt namespace
For some reason, after entering the mnt namespace of a running container
the root directory reported is the ChromeOS root, so the commands are not
limited to only seeing the container's rootfs, but they even manage to
escape the Crouton chroot.

The problem might be related to running runc in chroot, not to ChromeOS
or Crouton.

This patch is only intended for running runc on ChromeOS (in a chroot
managed by Crouton).