tests: Make sure we can traverse the path
Tests running in userns need to traverse the path for the rootfs as
different users (the one used in the userns mapping). This patch makes
sure all components in the path are always possible to traverse as
others, so the path should be ready to use on any test (using userns or
not).

We did something similar in 8542322dfe80b92c94291fd7427a9ca307017403
("libcontainer: Add unit tests with userns and mounts"), where we
traverse the path and make sure others always can read+exec. However,
here we were only changing the one component that didn't have the proper
permissions.

That worked fine in all distros, but breaks when upgrading to fedora 35.
So, let's just make sure the complete path has the proper permissions
and be done with it.

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>