[1.0] Fix failure with rw bind mount of a ro fuse
As reported in [1], in a case where read-only fuse (sshfs) mount
is used as a volume without specifying ro flag, the kernel fails
to remount it (when adding various flags such as nosuid and nodev),
returning EPERM.

Here's the relevant strace line:

> [pid 333966] mount("/tmp/bats-run-PRVfWc/runc.RbNv8g/bundle/mnt", "/proc/self/fd/7", 0xc0001e9164, MS_NOSUID|MS_NODEV|MS_REMOUNT|MS_BIND|MS_REC, NULL) = -1 EPERM (Operation not permitted)

I was not able to reproduce it with other read-only mounts as the source
(tried tmpfs, read-only bind mount, and an ext2 mount), so somehow this
might be specific to fuse.

The fix is to check whether the source has RDONLY flag, and retry the
remount with this flag added.

A test case (which was kind of hard to write) is added, and it fails
without the fix. Note that rootless user need to be able to ssh to
rootless@localhost in order to sshfs to work -- amend setup scripts
to make it work, and skip the test if the setup is not working.

[1] https://github.com/containers/podman/issues/12205

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
(cherry picked from commit 50105de1d87ed0994c32cc9fd57efd0a0f6ef544)

Conflicts:
 - .cirrus.yml: trivial, due to missing commit f0dbefac.
 - .github/workflows/test.yml: due to missing commits 120f74060 and
   3fd1851ce93e2, resolved manually.
 - Dockerfile: trivial, due to missing commit 24d318b8bb7a.
 - libcontainer/rootfs_linux.go: due to missing commits 36aefad45dd
   and 9c444070ec7bb8, resolved manually.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>