[1.0] ci: add Go 1.17, drop Go 1.15
Add Go 1.17, grop 1.15 from CI; since 1.17 release 1.15 is unsupported.

Keep 1.13 in 1.0 branch, since an older version of Docker/Moby might use
it.

Keep 1.16 in Dockerfile, since this is a stable branch and we'd rather
not swap horses in the middle of the stream.

This corresponds to commit a587180136b308222c in main branch.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>