libct: rm intelrtd.Manager interface, NewIntelRdtManager
Remove intelrtd.Manager interface, since we only have a single
implementation, and do not expect another one, remove the Manager
interface, rename intelRdtManager to Manager, and modify its users
accordingly.

Remove NewIntelRdtManager, as we only have a single implementation of
intelrdt.Manager, so there's nothing to configure here.

Remove IntelRdtfs as well. Instead, make intelrdt.NewManager
return nil if the feature is not available.

Remove TestFactoryNewIntelRdt as it is now a copy of TestFactoryNew.

Add internal function newManager to be used for tests (to make sure
some testing is done even when the feature is not available in
kernel/hardware).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>