runc: remove --criu option
This was introduced in an initial commit, and is not really needed for
most of the users, since they rely on CRIU binary provided by their
distro.

Now, if someone needs a custom criu binary to be used, they can put it
into some directory and make sure this directory is listed first in
$PATH.

Make --criu a hidden option (thus removing it from help). Remove it from
man pages, integration tests, etc. Remove all traces of CriuPath from
data structures.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>